requirements.txt
django==4.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
mssql-django==1.4.1
pyodbc==5.1.0